+++ src/libint/g_wxyz.h
@@ -4,9 +4,10 @@ extern "C" {
 #else
 #endif
 
-void construct_ao_g_wxyz_epsilon(double *g, long *s1, long *s2, long *s3, long *s4, double *epsilon, long *thread, long *skip,
-							long *n1, long *n2, long *n3, long *n4);
-void construct_ao_g_wxyz(double *g, long *s1, long *s2, long *s3, long *s4);
+void construct_ao_g_wxyz_epsilon(double *g, int *s1, int *s2, int *s3, int *s4, double *epsilon, 
+                                 int *thread, int *skip, int *n1, int *n2, int *n3, int *n4);
+
+void construct_ao_g_wxyz(double *g, int *s1, int *s2, int *s3, int *s4);
 
 #ifdef __cplusplus
 }
