+++ src/libint/h_wx.h
@@ -4,7 +4,7 @@ extern "C" {
 #else
 #endif
 
-void construct_ao_h_wx(double *h, long *s1, long *s2);
+void construct_ao_h_wx(double *h, int *s1, int *s2);
 
 #ifdef __cplusplus
 }
