+++ src/libint/h_wx_cdef.F90
@@ -7,7 +7,7 @@ interface
       implicit none
 !
       real(c_double), dimension(1,1) :: h
-      integer(c_long) :: s1, s2
+      integer(c_int) :: s1, s2
 !
    end subroutine construct_ao_h_wx_c
 !
