+++ src/libint/mu_wx.h
@@ -4,7 +4,7 @@ extern "C" {
 #else
 #endif
 
-void construct_ao_mu_wx(double *mu_X, double *mu_Y, double *mu_Z, long *s1, long *s2);
+void construct_ao_mu_wx(double *mu_X, double *mu_Y, double *mu_Z, int *s1, int *s2);
 
 #ifdef __cplusplus
 }
